Add .dockerignore file
Local container builds were breaking due to the contents of
the node_modules directory. This change adds a dockerignore
file, which prevents them from being copied to the container.

We also add a handful of other directories which could potentially
cause problems in future, or which are unnecessary to the app running.