Update Rack
Rack updates were resulting in CSRF errors on
attempting to log in. This was a result of the
fix for CVE-2020-8184 whic results in the cookie
keys no-longer being decoded. This was resulting
in the session cookie to be under 'SOME+SECRET'
rather than 'SOME SECRET'.

I've changed the session store back the the default.
The value isn't a secret, for one we expose it in
our source, but the key is trivially visible to anyone
who inspects their cookies.

This does have the unfortunate side effect of expiring
any sessions. However we can't simply modify the
key to 'SOME+SECRET' as that itself is escaped to
'SOME%2BSECRET'.