Remove valid? from null context
valid? here returns a boolean, which itself
doesn't actually do anything in this context.
While we *do* populate the errors object,
we don't use this in any context, without
actually caling valid? first. This re-runs
the validation and obliterates the orignal
errors.

This action is surprisingly expensive,
and, although we hope to make it cheaper,
we don't want to be running this operation
needelessly.