Switch to yard for documentation
We generally use yard for our documentation, plus it works better
with the ruby language servers like solrgraph.

Sdoc was also causing frustrating dependency issues, as it required
rdoc as a dependency, which in turn imported psych as a gem. This
resulted in breaking changes (in a minor version, as they don't follow
semver) when rdoc suddenly depended on psych 4 which changes the way
yaml gets loaded by default. Rather than pinning psych, I've switched
to yard.