Fix tube import
The tube import tests weren't explicitly testing the facts
and were improperly mocked. We switch to webmock to ensure
the tested resource accurately reflect those seen in reality, and
explicitly test the added facts.

These tests revealed that we no longer responded to aliquots,
so didn't import the data for tubes. This adds a compatible API,
although in practice the import handling could do with further tidy-up.

This change also reveals that the initial facts are immediately
cleared away. This use of separate import and refresh steps seems
a little unnecessary, and I'll be unifying them in the next commit.