Unify import and refresh steps
On importing an asset were were creating an import step that was never
properly resolved, and a refresh step. The latter overrode the facts
associated with the former, and yet was somewhat misleading, as this
all occurred on the first import.

This change removes the redundant facts and passes the import step
in to the refresh process, such that it passes correctly.