Fix refresh of well modeled tubes
Fixes DPL-nnn [BUG] Reracking failing on every alternative run #205

This was a tricky one to narrow down.
Once a tube is exported to
Sequencescape it becomes modeled
as a well. On subsequent CSV
upload, we were refreshing the
well information, but the import
action was not adding the location data.

This fix solves the immediate issue, but is a little incomplete
as is still wipes out parents.